qa: move recursive signature warmup to dedicated class
