+++ src/MapperBuilder.php
@@ -9,11 +9,11 @@ use CuyZ\Valinor\Library\Container;
 use CuyZ\Valinor\Library\Settings;
 use CuyZ\Valinor\Mapper\Tree\Node;
 use CuyZ\Valinor\Mapper\TreeMapper;
-use CuyZ\Valinor\Type\CompositeType;
 use CuyZ\Valinor\Type\Parser\Exception\InvalidType;
 use CuyZ\Valinor\Type\Types\ClassType;
 use Psr\SimpleCache\CacheInterface;
 
+use CuyZ\Valinor\Utility\RecursiveSignatureTypeDefinitionWarmup;
 use function array_merge;
 use function is_callable;
 
@@ -291,33 +291,9 @@ final class MapperBuilder
     public function warmup(string $signature, string ...$additionalSignatures): void
     {
         $signaturesToWarmup = array_merge([$signature], $additionalSignatures);
-        $this->warmupSignatures(array_values($signaturesToWarmup), []);
-    }
-
-    /**
-     * @param list<class-string> $signatures
-     * @param list<class-string> $alreadyKnownSignatures
-     * @return list<class-string>
-     * @throws InvalidType In case one of the provided signatures contain invalid types.
-     */
-    private function warmupSignatures(array $signatures, array $alreadyKnownSignatures): array
-    {
-        $signaturesWarmedUp = $alreadyKnownSignatures;
         $container = $this->container();
-        $typeParser = $container->typeParser();
-
-        foreach ($signatures as $signature) {
-            if (in_array($signature, $signaturesWarmedUp, true)) {
-                continue;
-            }
-
-            $type = $typeParser->parse($signature);
-            $signaturesWarmedUp[] = $signature;
-
-            $signaturesWarmedUp = $this->warmupSignatureByType($type, $signaturesWarmedUp);
-        }
-
-        return $signaturesWarmedUp;
+        (new RecursiveSignatureTypeDefinitionWarmup($container->typeParser(), $container->classDefinitionRepository()))
+            ->warmupSignatures(array_values($signaturesToWarmup), []);
     }
 
     public function __clone()
@@ -329,58 +305,5 @@ final class MapperBuilder
     private function container(): Container
     {
         return ($this->container ??= new Container($this->settings));
-    }
-
-    /**
-     * @param list<class-string> $alreadyKnownSignatures
-     * @return list<class-string>
-     */
-    private function extractSignaturesFromClassDefinition(Definition\ClassDefinition $classDefinition, array $alreadyKnownSignatures): array
-    {
-        $recursiveSignatures = [];
-        foreach ($classDefinition->properties() as $property) {
-            $propertyType = $property->type();
-            if (!$propertyType instanceof ClassType) {
-                continue;
-            }
-
-            $className = $propertyType->className();
-            if (in_array($className, $alreadyKnownSignatures, true)) {
-                continue;
-            }
-
-            $recursiveSignatures[] = $className;
-        }
-
-        return $recursiveSignatures;
-    }
-
-    /**
-     * @param list<class-string> $alreadyKnownSignatures
-     *
-     * @return list<class-string>
-     * @throws InvalidType In case one of the provided signatures contain invalid types.
-     */
-    private function warmupSignatureByType(Type\Type $type, array $alreadyKnownSignatures): array
-    {
-        if ($type instanceof ClassType) {
-            $classDefinitionRepository = $this->container()->classDefinitionRepository();
-            $classDefinition = $classDefinitionRepository->for($type);
-            $recursiveSignaturesToWarmup = $this->extractSignaturesFromClassDefinition($classDefinition, $alreadyKnownSignatures);
-
-            return $this->warmupSignatures($recursiveSignaturesToWarmup, $alreadyKnownSignatures);
-        }
-
-        if ($type instanceof CompositeType) {
-            $types = $type;
-            unset($type);
-            foreach ($types->traverse() as $type) {
-                $alreadyKnownSignatures = $this->warmupSignatureByType($type, $alreadyKnownSignatures);
-            }
-
-            return $alreadyKnownSignatures;
-        }
-
-        return $alreadyKnownSignatures;
     }
 }
